enable tests on ScalaJS
