+++ streams-tests/shared/src/test/scala/zio/stream/StreamHaltWhenSpec.scala
