+++ streams-tests/shared/src/test/scala/zio/stream/StreamInterruptWhenSpec.scala
